Moved tests to correct location Refactored some tests
