+++ tests/cli/Cli/Dispatcher/DispatchCest.php
@@ -15,17 +15,10 @@ namespace Phalcon\Test\Cli\Cli\Dispatcher;
 use CliTester;
 use Phalcon\Cli\Dispatcher;
 use Phalcon\Cli\Dispatcher\Exception;
-use Phalcon\Test\Fixtures\Traits\DiTrait;
+use Phalcon\Di\FactoryDefault\Cli as DiFactoryDefault;
 
 class DispatchCest
 {
-    use DiTrait;
-
-    public function _before(CliTester $I)
-    {
-        $this->setNewCliFactoryDefault();
-    }
-
     /**
      * Tests Phalcon\Cli\Dispatcher :: dispatch()
      *
@@ -36,7 +29,139 @@ class DispatchCest
     {
         $I->wantToTest('Cli\Dispatcher - dispatch()');
 
-        $I->skipTest('Need implementation');
+        // test 1
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDI(
+            new DiFactoryDefault()
+        );
+        $dispatcher->setDefaultNamespace('Phalcon\Test\Fixtures\Tasks');
+        $dispatcher->dispatch();
+        $I->assertEquals(
+            'main',
+            $dispatcher->getTaskName()
+        );
+        $I->assertEquals(
+            'main',
+            $dispatcher->getActionName()
+        );
+        $I->assertEquals(
+            [],
+            $dispatcher->getParams()
+        );
+        $I->assertEquals(
+            'mainAction',
+            $dispatcher->getReturnedValue()
+        );
+
+
+        // Test 2
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDI(
+            new DiFactoryDefault()
+        );
+        $dispatcher->setDefaultNamespace('Phalcon\Test\Fixtures\Tasks');
+        $dispatcher->setTaskName('echo');
+        $dispatcher->dispatch();
+        $I->assertEquals(
+            'echo',
+            $dispatcher->getTaskName()
+        );
+        $I->assertEquals(
+            'main',
+            $dispatcher->getActionName()
+        );
+        $I->assertEquals(
+            [],
+            $dispatcher->getParams()
+        );
+        $I->assertEquals(
+            'echoMainAction',
+            $dispatcher->getReturnedValue()
+        );
+
+        // Test 3
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDI(
+            new DiFactoryDefault()
+        );
+        $dispatcher->setDefaultNamespace('Phalcon\Test\Fixtures\Tasks');
+        $dispatcher->setTaskName('main');
+        $dispatcher->setActionName('hello');
+        $dispatcher->dispatch();
+        $I->assertEquals(
+            'main',
+            $dispatcher->getTaskName()
+        );
+        $I->assertEquals(
+            'hello',
+            $dispatcher->getActionName()
+        );
+        $I->assertEquals(
+            [],
+            $dispatcher->getParams()
+        );
+        $I->assertEquals(
+            'Hello !',
+            $dispatcher->getReturnedValue()
+        );
+
+        // Test 4
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDI(
+            new DiFactoryDefault()
+        );
+        $dispatcher->setDefaultNamespace('Phalcon\Test\Fixtures\Tasks');
+        $dispatcher->setActionName('hello');
+        $dispatcher->setParams(
+            [
+                'World',
+                '######',
+            ]
+        );
+        $dispatcher->dispatch();
+        $I->assertEquals(
+            'main',
+            $dispatcher->getTaskName()
+        );
+        $I->assertEquals(
+            'hello',
+            $dispatcher->getActionName()
+        );
+        $I->assertEquals(
+            [
+                'World',
+                '######',
+            ],
+            $dispatcher->getParams()
+        );
+        $I->assertEquals(
+            'Hello World######',
+            $dispatcher->getReturnedValue()
+        );
+
+        //Test 5
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDI(
+            new DiFactoryDefault()
+        );
+        $dispatcher->setDefaultNamespace('Phalcon\Test\Fixtures\Tasks');
+        $dispatcher->setActionName('hello');
+        $dispatcher->setParams(
+            [
+                'hello'   => 'World',
+                'goodbye' => 'Everybody',
+            ]
+        );
+        $dispatcher->dispatch();
+        $I->assertTrue(
+            $dispatcher->hasParam('hello')
+        );
+        $I->assertTrue(
+            $dispatcher->hasParam('goodbye')
+        );
+        $I->assertFalse(
+            $dispatcher->hasParam('salutations')
+        );
     }
 
     public function testFakeNamespace(CliTester $I)
@@ -44,7 +169,7 @@ class DispatchCest
         $dispatcher = new Dispatcher();
 
         $dispatcher->setDI(
-            $this->container
+            new DiFactoryDefault()
         );
 
         $dispatcher->setDefaultNamespace('Dummy\\');
